- Added test for currency helper

+++ app/src/main/java/io/jonibek/currency/data/remote/json/BaseResult.kt
@@ -1,7 +1,8 @@
 package io.jonibek.currency.data.remote.json
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class BaseResult(@SerializedName("base") var base: String,
                       @SerializedName("date") var date: String,
-                      @SerializedName("rates") var rates : Map<String,Float>)
+                      @SerializedName("rates") var rates : Map<String,BigDecimal>)
